+++ photutils/psf/core.py
@@ -596,10 +596,22 @@ def prepare_psf_model(psfmodel, xname=None, yname=None, fluxname=None,
 @support_nddata
 def psf_photometry(data, positions, psf, fitshape=None,
                    fitter=LevMarLSQFitter(),
+<<<<<<< HEAD:photutils/psf/core.py
+<<<<<<< HEAD:photutils/psf/core.py
                    unit=None, wcs=None, error=None,
                    mask=None, pixelwise_error=True,
                    mode='sequential',
                    store_fit_info=False):
+=======
+                   unit=None, wcs=None, error=None, effective_gain=None,
+                   mask=None, pixelwise_error=True, mode='sequential',
+                   store_fit_info=False, param_uncert=False):
+>>>>>>> add changelog and proper warning in case fitter does not contain 'param_cov' key:photutils/psf.py
+=======
+                   unit=None, wcs=None, error=None, effective_gain=None,
+                   mask=None, pixelwise_error=True, mode='sequential',
+                   store_fit_info=False, param_uncert=False):
+>>>>>>> a71068eeed44954dde0cf693976dc926a3b60bbe:photutils/psf.py
     """
     Perform PSF/PRF photometry on the data.
 
@@ -679,6 +691,11 @@ def psf_photometry(data, positions, psf, fitshape=None,
         table will have an additional column 'fit_message' with the message that
         came from the fit.  If a list, it will be populated with the
         ``fit_info`` dictionary of the fitter for each fit.
+    param_uncert : bool (default=False)
+        If True, the uncertainties on each parameter estimate will be stored
+        in the output table. This option assumes that the fitter has the
+        'param_cov' key in its 'fit_info' dictionary.
+        See 'fit_info' in `~astropy.modeling.fitting.LevMarLSQFitter`.
 
     Returns
     -------
@@ -767,6 +784,15 @@ def psf_photometry(data, positions, psf, fitshape=None,
         fit_infos = store_fit_info
     elif store_fit_info:
         fit_messages = []
+    if param_uncert:
+        if 'param_cov' in fitter.fit_info:
+            uncert = []
+        else:
+            warnings.warn("uncertainties on fitted parameters cannot be " +
+                           "computed because fitter does not contain " +
+                           "`param_cov` key in its `fit_info` dictionary.",
+                           AstropyUserWarning)
+            param_uncert = False
 
     # Many fitters take a "weight" array, but no "mask".
     # Thus, we convert the mask to weights on 1 and 0. Unfortunately,
@@ -802,11 +828,28 @@ def psf_photometry(data, positions, psf, fitshape=None,
                 fit_infos.append(fitter.fit_info)
             if fit_messages is not None:
                 fit_messages.append(fitter.fit_info['message'])
+            if param_uncert:
+                if fitter.fit_info['param_cov'] is not None:
+                    uncert.append(np.sqrt(np.diag(\
+                            fitter.fit_info['param_cov'])))
+                else:
+                    warnings.warn("uncertainties on fitted parameters " +
+                                  "cannot be computed because the fit may " +
+                                  "be unsuccessful", AstropyUserWarning)
+                    uncert.append((None, None, None))
     else:
         raise ValueError('Invalid photometry mode.')
 
     if fit_messages is not None:
         result_tab['fit_messages'] = fit_messages
+    if param_uncert:
+        uncert = np.array(uncert)
+        i = 0
+        for param in psf.param_names:
+            if not getattr(psf, param).fixed:
+                result_tab.add_column(Column(name=param + "_fit_uncertainty",
+                                             unit=None, data=uncert[:, i]))
+                i += 1
 
     return result_tab