+++ CHANGES.rst
@@ -25,6 +25,8 @@ New Features
 
   - Added new ``make_source_mask`` convenience function. [#355]
 
+<<<<<<< HEAD
+<<<<<<< HEAD
 - ``photutils.segmentation``
 
   - Added ``copy`` and ``area`` methods and a ``areas`` property to
@@ -33,6 +35,17 @@ New Features
   - SegmentationImage properties are now cached to significantly
     improve performance. [#361]
 
+=======
+=======
+>>>>>>> a71068eeed44954dde0cf693976dc926a3b60bbe
+- ``photutils.psf_photometry``
+
+  - Added the option ``param_uncert`` so that users can easily get
+    uncertainties on fitted parameters. [#358]
+<<<<<<< HEAD
+>>>>>>> add changelog and proper warning in case fitter does not contain 'param_cov' key
+=======
+>>>>>>> a71068eeed44954dde0cf693976dc926a3b60bbe
 
 API changes
 ^^^^^^^^^^^